Fixed logic for dropping worst sample
I believe that delta_obj_continuous as computed is the bin-count for each histogram summed up across histograms. If the first bin is the furthest from eta it makes no sense to drop the first sample. The previous commit updates the logic to track contributions from each sample to each bin. Sample.weights are used not to evaluated which is worst before dropping.