Updated clhs.data.frame.R to fix names bug
R data frames can handle non-standard parameters now. And GIS layers seem to be full of them. My recent experience with one labelled "DSM_GEOTIFF/1" encountered a problem with this line where the creation of the data.frame on this line changed the name to "DSM_GEOTIFF.1" which causes a problem when we reorder the cols on like 274 as one of the names now doesn't exist.